Add lazy package for types to be resolved at runtime
This package is for function type definitions that return some value, for
example a function that takes a pointer to a Playbook and returns a string
value. These are used to resolve values in tasks at runtime, for example a
`lazy.String` can return a string value from the output of a task which
has previously run.

A `Frame` field has been added to `Playbook` to carry state and variables
between tasks. The frame is passed to all lazy functions so they are able
to look up values from previous tasks.