Add the ability for a task to retrieve the outputs of another
This allows a task to get the output of another task as input. A lazy
`Sprintf()` has been added which, instead of having a variadic argument
of interfaces, has a variadic argument of "lazy" interfaces which are
functions that receive the playbook `Frame` and return an interface. Using
this, it is possible to look up the outputs of tasks which have already run.