Modify lazy functions to return zero-argument inner function
Where a lazy String function was previously defined (for example):

func S(s string) String {
        return func(*types.Frame) string {
                return s
        }
}

It is now defined as:

func S(s string) String {
        return func(*types.Frame) StringValue {
                return func() string {
                        return s
                }
        }
}

and where StringValue is a `func() string`.

Now, instead of having modules pass the Frame to the lazy String, it is
done in the playbook. The module only needs to call the StringValue function.

Modules will define all public fields as StringValue, BoolValue, MapValue, etc.