Make all lazy functions return an error value
This enables a module to return an error at runtime instead of silently
swallowing it, for example in cases where a task does a lookup into a previous
task's output and it is not found. Instead of returning an empty string as a
zero value, it will now return an error and the module can report it.