Add JSON input file with validation using JSON schema
This adds an InputSchema field to playbooks, which must define a JSON schema
as a map[string]interface{}. It also adds a --vars CLI option to the apply
subcommand of a playbook, which must be valid JSON according to the defined
schema.

To make use of this new feature, the CloudFormation module has been changed
to take an optional TemplateFile. This is the path to the CloudFormation
template as a file. For reference, the example playbook at the root of the
repository has been changed to use this, as well as having a vars.json
which defines the path to the template file.