Add the ability to compile a YAML playbook to a binary
This adds a compiler as library functions `Load()` and `Compile()`, which can
load YAML and generate a Go AST. The AST can be formatted into a Go file and
compiled into a binary.

As yet there is no CLI compiler, only the library functions.

This includes an example playbook.yml at the root of the repository which
showcases the syntax and features available in a playbook.