Instantiate modules lazily
This changes the `task.Task` to have a function field `Unwrap` that, when
called, returns the module or an error. The playbook calls this `Unwrap`
function as it iterates through the tasks while running. If there is an error,
the playbook is aborted, otherwise the returned module is assigned to the task's
`Module` field.

By making this change, modules no longer need to be aware of special types defined
in the `lazy` package, such as `lazy.String`. Instead the module fields can be
ordinary types such as string.