Switch the compiler to use the new syntax
The generated PEG parser creates an AST when parsing an Unobin playbook. This
is traversed to create an Unobin AST (UAST). The UAST is validated to ensure
the playbook has all required attributes of the correct types, and that tasks
only refer to modules which have been imported. UAST value types have a
ToGoAST() method, which is used in generating the Go AST from the UAST.

The CLI `init` command has been changed to generate "playbook.ub" using the
new syntax instead of "playbook.yml".