Redesign builtin playbook functions
Previously, a builtin playbook function such as `bool-var` worked as follows:

func BoolVar(path StringValue) Bool {
	return func(ctx *types.Context) BoolValue {
		return func() (bool, error) {
			// function body omitted
		}
	}
}

With the new design, the same function is defined as:

func BoolVar(ctx *types.Context, path String) Bool {
	// Error of arguments need to be checked first.
	if path.Error != nil {
		return Bool{Error: path.Error}
	}
	// function body omitted
	return Bool{Value: true, Error: nil}
}

Bool and other supported types are defined as a struct containing both the
value and error:

type Bool struct {
	Value bool
	Error error
}

Since the new functions are not lazy as in the previous version, they are
wrapped in a closure at compile time in the playbook, which makes functions
simpler to implement.

Wrapping types in a struct containing the value and the error allows builtin
functions to compose easily. All builtin functions must receive and return
only these types, aside from the initial Context argument. All builtin
functions must check for errors of such arguments before proceeding to the
main body of the function.