Add a template module and the ability to bundle resources
For the template module to work, there must be a way to bundle templates into
the playbook binary. To achieve this, github.com/go-bindata/go-bindata has been
embedded into the compiler, and any files in the resources subdirectory of a
playbook will be bundled into a file called resources.go. When compiling a
playbook, the user can pass `--skip-resources` to skip the bundling step. When
a playbook runs, the bundled files will be written to a cache directory,
defaulting to ~/.unobin/<playbook-name>.