Make generated Go easier to read
This changes Go code to be generated with github.com/dave/dst instead of
go/ast, as it has "decorations" added to nodes that enable the addition
of whitespace. Now structs and maps have newlines added before and after
opening and closing braces, which makes the code more human readable.