Display output from Go compiler
The first time compiling a playbook can take time as Go dependencies are
downloaded. It can give an impression of the compiler being slow if the
output of the Go compiler is not shown. This changes it so the Go compiler
output is displayed, including showing the downloading of dependencies.