Enhance module arguments to enable functions within collection
It has been possible for module arguments to be function calls, but not if
the function call is within a collection (object or array). For example, it
was possible to do:

task [expand a template] {
  module: template
  args: {
    src: string-var('path')    # Function possible here.
    vars: { 'name': 'Unobin' } # Function not possible within object.
  }
}

With this commit, it is now possible for the object to contain function calls:

task [expand a template] {
  module: template
  args: {
    src: 'thing.conf.tmpl'
    vars: { 'name': string-var('name') } # Function now possible here.
  }
}

Function calls can be at any nested level within a collection.