Declare a variable to prevent a compile error in playbooks
The package `github.com/cloudboss/unobin/pkg/functions` is always imported into
all playbooks, which caused a compile error if no functions were used in a
playbook. This declares a variable `_` of type `functions.Array` in the main
function of a playbook, so that there will not be a compile error.