Fix panic when playbook is missing an input schema
The validation step of the compile collects as many failures as possible
before returning, but in this case it was trying to validate the input schema
even if it was earlier found to be missing. This puts a conditional to skip
the input schema validation if it was not found.